latest work to work with eapi72 kit-fixups branch. Appears working now.

+++ targets/gentoo/steps/unpack.spec
@@ -86,18 +86,21 @@ EOF
 	for mixin in $mixins; do
 		echo $mixin >> $[path/chroot]/etc/portage/make.profile/parent
 	done
-	cat > $[path/chroot]/etc/ego.conf << EOF
+	if [ -e /etc/metro/chroot/etc/ego.conf ]; then
+		cp /etc/metro/chroot/etc/ego.conf $[path/chroot]/etc/ego.conf
+	else 
+		cat > $[path/chroot]/etc/ego.conf << EOF
 [global]
 sync_base_url = $[snapshot/source/sync_base_url]
 EOF
-	if [ "$[snapshot/source/ego.conf?]" = "yes" ]
-	then
-		echo "Installing /etc/ego.conf..."
-		cat >> $[path/chroot]/etc/ego.conf << EOF
+		if [ "$[snapshot/source/ego.conf?]" = "yes" ]
+		then
+			echo "Installing /etc/ego.conf..."
+			cat >> $[path/chroot]/etc/ego.conf << EOF
 $[[snapshot/source/ego.conf]]
 EOF
+		fi
 	fi
-
 
 		cat $[path/chroot]/etc/ego.conf
 		ROOT=$[path/chroot] /root/ego/ego sync --kits-only || exit 8
